requirements.txt
fastapi[standard]
uvicorn
pydantic
python-jose
passlib
pymongo
python-multipart